+++ sdk/spring/azure-spring-boot/CHANGELOG.md
@@ -13,6 +13,10 @@
           user-flows:
             ${your-user-flow-key}: ${your-user-flow-name-defined-on-azure-portal}
     ```
+- Require new property of `spring.jms.servicebus.pricing-tier` to set pricing tier of Azure Service Bus. Supported values are `premium`, `standard` and `basic`.
+### New Features
+- Enable MessageConverter bean customization.
+- Update the underpinning JMS library for the Premium pricing tier of Service Bus to JMS 2.0.
 
 ## 3.1.0 (2021-01-20)
 ### Breaking Changes